Added goroutines for better performance

+++ scraper-worker/internal/models.go/models.go
@@ -13,7 +13,7 @@ type Task struct {
 
 type ScrapedResult struct {
 	JobID   int64  `json:"job_id"`
-	TaskID  int64  `json:"task_id"`
+	TaskID  int    `json:"task_id"`
 	URL     string `json:"url"`
 	RawText string `json:"raw_text"`
 	Error   string `json:"error,omitempty"`
